adjusted sorting, added logs and minor improvements

+++ bin/handler.js
@@ -1,9 +1,10 @@
 import { join } from "path"
-import { readFileSync, writeFileSync } from "fs"
+import { existsSync, readFileSync, writeFileSync } from "fs"
 import sortPackage from "../src/index.js"
 
 export default function handler ( folder = "." ) {
 	const pkgPath = join(process.cwd(), folder, "package.json")
+	if (!existsSync(pkgPath)) return console.log(`${join(folder, "package.json").replace(/\\/g, "/")} does not exist.`)
 
 	writeFileSync(
 		pkgPath,
@@ -11,4 +12,6 @@ export default function handler ( folder = "." ) {
 			readFileSync(pkgPath)
 		)
 	)
+
+	console.log(`sorted ${join(folder, "package.json").replace(/\\/g, "/")}`)
 }
